# nvim-treesitter: use buildEnv to reduce rtp pollution