@PerchunPak 2025-05-18
Is it equal to https://github.com/NixOS/nixpkgs/pull/319233?