@Gerg-L 2025-05-18
So the parsers and everything with tree-sitter works with this. but it still has the issues with dependencies (attempting to be fixed in #319233)